refactor(routing): inline private resizer support types to enable type re-export
PR #1584 review (cursor, Low): `OptimalSizeExploringResizer` was reachable
only via the verbose path `routing::optimal_size_exploring_resizer::OptimalSizeExploringResizer`
while every other public type in the module (`DefaultResizer`, `PoolRouter`,
`Resizer`, `Routers`, …) is re-exported under `routing::`. Fix the
inconsistency.

The prior reason for the verbose path was that `no-parent-reexport`
(module_wiring_lint) forbids `pub use child::Type;` when `child`
itself has sub-modules, and `optimal_size_exploring_resizer/` had four
`pub(crate)` support modules (`lcg`, `state`, `resize_record`,
`under_utilization_streak`). Relaxing that lint would violate the
project-wide "do not silence lints with allow" rule.

Resolution: inline the four small (~14..47 line) support types into
`optimal_size_exploring_resizer.rs`. The only remaining child is the
`#[cfg(test)] mod tests;`, which disappears from HIR in non-test
builds, so the module now passes leaf detection and `pub use
optimal_size_exploring_resizer::OptimalSizeExploringResizer;` is legal.

Users can now write `use ...::routing::OptimalSizeExploringResizer`
matching the rest of the routing DSL.

No behavior change. All 1716 lib tests pass; `./scripts/ci-check.sh ai
all` exit 0.
